DRY up device movement logic.
