+++ nougat/utils/device.py
@@ -1,0 +1,8 @@
+import torch
+
+def move_to_device(model):
+    if torch.cuda.is_available():
+        return model.to("cuda")
+    elif torch.backends.mps.is_available():
+        return model.to("mps")
+    return model
